Add a flag for overriding the runtime dir.
Adds a --runtime-dir cli flag which overrides the /run dir in the lxcfslib. This ended up being kind of tricky because of how lxcfslib can be reloaded and its use of a library constructor. In order read the cli flag and then set a variable in the library, I removed the contstructor attribute and made it an explicit call in order to override the runtime dir in between loading the library and initializing lxcfslib.

Signed-off-by: Sebastien Dabdoub <sebastien.dabdoub@gmail.com>